requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
pdfminer.six
PyPDF2
pandas
markdown
pygments
numpy
langchain-community
tavily-python==0.2.2
google-search-results
